requirements.txt
colorgram.py==1.1.0
Flask==1.0.2
requests==2.18.4
Pillow==5.1.0
PyYAML==3.12
